backend/requirements.txt
django-cors-headers==3.5.0
django-extra-fields==3.0.1